Backend/requirements.txt
boto3
strands-agents
bedrock-agentcore
regex
bedrock-agentcore<=0.1.5
bedrock-agentcore-starter-toolkit==0.1.14
